+++ pkgs/dart_tooling_mcp_server/test/tools/dtd_test.dart
@@ -4,8 +4,10 @@
 
 import 'dart:convert';
 
+import 'package:async/async.dart';
 import 'package:dart_mcp/server.dart';
 import 'package:dart_tooling_mcp_server/src/mixins/dtd.dart';
+import 'package:dart_tooling_mcp_server/src/utils/constants.dart';
 import 'package:test/test.dart';
 import 'package:vm_service/vm_service.dart';
 
@@ -61,65 +63,6 @@ void main() {
         expect(hotReloadResult.content, [
           TextContent(text: 'Hot reload succeeded.'),
         ]);
-      });
-
-      test('can get runtime errors', () async {
-        await testHarness.startDebugSession(
-          counterAppPath,
-          'lib/main.dart',
-          isFlutter: true,
-          args: ['--dart-define=include_layout_error=true'],
-        );
-        final tools = (await testHarness.mcpServerConnection.listTools()).tools;
-        final runtimeErrorsTool = tools.singleWhere(
-          (t) => t.name == DartToolingDaemonSupport.getRuntimeErrorsTool.name,
-        );
-        final runtimeErrorsResult = await testHarness.callToolWithRetry(
-          CallToolRequest(name: runtimeErrorsTool.name),
-        );
-
-        expect(runtimeErrorsResult.isError, isNot(true));
-        final errorCountRegex = RegExp(r'Found \d+ errors?:');
-        expect(
-          (runtimeErrorsResult.content.first as TextContent).text,
-          contains(errorCountRegex),
-        );
-        expect(
-          (runtimeErrorsResult.content[1] as TextContent).text,
-          contains('A RenderFlex overflowed by'),
-        );
-
-        final now = DateTime.now().millisecondsSinceEpoch;
-        final sinceNowResult = await testHarness.callToolWithRetry(
-          CallToolRequest(
-            name: runtimeErrorsTool.name,
-            arguments: {'since': now},
-          ),
-        );
-        expect(
-          (sinceNowResult.content.first as TextContent).text,
-          contains('No runtime errors found'),
-        );
-
-        // Trigger a hot reload, should see the error again.
-        await testHarness.callToolWithRetry(
-          CallToolRequest(name: DartToolingDaemonSupport.hotReloadTool.name),
-        );
-
-        final finalRuntimeErrorsResult = await testHarness.callToolWithRetry(
-          CallToolRequest(
-            name: runtimeErrorsTool.name,
-            arguments: {'since': now},
-          ),
-        );
-        expect(
-          (finalRuntimeErrorsResult.content.first as TextContent).text,
-          contains(errorCountRegex),
-        );
-        expect(
-          (finalRuntimeErrorsResult.content[1] as TextContent).text,
-          contains('A RenderFlex overflowed by'),
-        );
       });
 
       test('can get the widget tree', () async {
@@ -230,7 +173,7 @@ void main() {
           final children = widgetTree['children'] as List<Object?>;
           final firstWidgetId =
               (children.first as Map<String, Object?>)['valueId'];
-          final appVmService = server.activeVmServices.values.first;
+          final appVmService = await server.activeVmServices.values.first;
           final vm = await appVmService.getVM();
           await appVmService.callServiceExtension(
             'ext.flutter.inspector.setSelectionById',
@@ -279,6 +222,178 @@ void main() {
           );
         });
       });
+
+      group('runtime errors', () {
+        final errorCountRegex = RegExp(r'Found \d+ errors?:');
+
+        setUp(() async {
+          await testHarness.startDebugSession(
+            counterAppPath,
+            'lib/main.dart',
+            isFlutter: true,
+            args: ['--dart-define=include_layout_error=true'],
+          );
+        });
+
+        test('can be read and cleared using the tool', () async {
+          final tools =
+              (await testHarness.mcpServerConnection.listTools()).tools;
+          final runtimeErrorsTool = tools.singleWhere(
+            (t) => t.name == DartToolingDaemonSupport.getRuntimeErrorsTool.name,
+          );
+          late CallToolResult runtimeErrorsResult;
+
+          // Give the errors at most a second to come through.
+          var count = 0;
+          while (true) {
+            runtimeErrorsResult = await testHarness.callToolWithRetry(
+              CallToolRequest(
+                name: runtimeErrorsTool.name,
+                arguments: {'clearRuntimeErrors': true},
+              ),
+            );
+            expect(runtimeErrorsResult.isError, isNot(true));
+            final firstText =
+                (runtimeErrorsResult.content.first as TextContent).text;
+            if (errorCountRegex.hasMatch(firstText)) {
+              break;
+            } else if (++count > 10) {
+              fail('No errors found, expected at least one');
+            } else {
+              await Future<void>.delayed(const Duration(milliseconds: 100));
+            }
+          }
+          expect(
+            (runtimeErrorsResult.content[1] as TextContent).text,
+            contains('A RenderFlex overflowed by'),
+          );
+
+          // We cleared the errors in the previous call, shouldn't see any here.
+          final nextResult = await testHarness.callToolWithRetry(
+            CallToolRequest(name: runtimeErrorsTool.name),
+          );
+          expect(
+            (nextResult.content.first as TextContent).text,
+            contains('No runtime errors found'),
+          );
+
+          // Trigger a hot reload, should see the error again.
+          await testHarness.callToolWithRetry(
+            CallToolRequest(name: DartToolingDaemonSupport.hotReloadTool.name),
+          );
+
+          final finalRuntimeErrorsResult = await testHarness.callToolWithRetry(
+            CallToolRequest(name: runtimeErrorsTool.name),
+          );
+          expect(
+            (finalRuntimeErrorsResult.content.first as TextContent).text,
+            contains(errorCountRegex),
+          );
+          expect(
+            (finalRuntimeErrorsResult.content[1] as TextContent).text,
+            contains('A RenderFlex overflowed by'),
+          );
+        });
+
+        test('can be read and subscribed to as a resource', () async {
+          final serverConnection = testHarness.mcpServerConnection;
+          final onResourceListChanged =
+              serverConnection.resourceListChanged.first;
+          var resources =
+              (await serverConnection.listResources(
+                ListResourcesRequest(),
+              )).resources;
+          if (resources.runtimeErrors.isEmpty) {
+            await onResourceListChanged;
+            resources =
+                (await serverConnection.listResources(
+                  ListResourcesRequest(),
+                )).resources;
+          }
+          final resource = resources.runtimeErrors.single;
+
+          final resourceUpdatedQueue = StreamQueue(
+            serverConnection.resourceUpdated,
+          );
+          await serverConnection.subscribeResource(
+            SubscribeRequest(uri: resource.uri),
+          );
+          await pumpEventQueue();
+          var originalContents =
+              (await serverConnection.readResource(
+                ReadResourceRequest(uri: resource.uri),
+              )).contents;
+          final overflowMatcher = isA<TextResourceContents>().having(
+            (c) => c.text,
+            'text',
+            contains('A RenderFlex overflowed by'),
+          );
+          // If we haven't seen errors initially, then listen for updates and
+          // re-read the resource.
+          if (originalContents.isEmpty) {
+            await resourceUpdatedQueue.next;
+            originalContents =
+                (await serverConnection.readResource(
+                  ReadResourceRequest(uri: resource.uri),
+                )).contents;
+          }
+          // Sometimes we get this error logged multiple times
+          expect(originalContents.first, overflowMatcher);
+          await pumpEventQueue();
+
+          await testHarness.callToolWithRetry(
+            CallToolRequest(name: DartToolingDaemonSupport.hotReloadTool.name),
+          );
+
+          expect(
+            await resourceUpdatedQueue.next,
+            isA<ResourceUpdatedNotification>().having(
+              (n) => n.uri,
+              ParameterNames.uri,
+              resource.uri,
+            ),
+          );
+
+          // We should see additional errors (but the exact number is variable).
+          final newContents =
+              (await serverConnection.readResource(
+                ReadResourceRequest(uri: resource.uri),
+              )).contents;
+          expect(newContents.length, greaterThan(originalContents.length));
+          expect(newContents.last, overflowMatcher);
+
+          // Now hot reload but clear previous errors, should see fewer errors
+          // than before after this.
+          await testHarness.callToolWithRetry(
+            CallToolRequest(
+              name: DartToolingDaemonSupport.hotReloadTool.name,
+              arguments: {'clearRuntimeErrors': true},
+            ),
+          );
+
+          expect(
+            await resourceUpdatedQueue.next,
+            isA<ResourceUpdatedNotification>().having(
+              (n) => n.uri,
+              ParameterNames.uri,
+              resource.uri,
+            ),
+          );
+
+          final finalContents =
+              (await serverConnection.readResource(
+                ReadResourceRequest(uri: resource.uri),
+              )).contents;
+          expect(finalContents.length, lessThan(newContents.length));
+          expect(finalContents.last, overflowMatcher);
+        });
+      });
     });
   });
+}
+
+extension on Iterable<Resource> {
+  Iterable<Resource> get runtimeErrors => where(
+    (r) => r.uri.startsWith(DartToolingDaemonSupport.runtimeErrorsScheme),
+  );
 }